guard bound-free rate calculation against DeltaEnergy <= 0
Guards the rate formulas for bound-free transitions against deltaEnergy <= 0.

Fixes the rate calculation formulas returning NaNs and/or +-infinity due to the energy difference between the lower and upper state of a bound free transition being <= 0 due to ionization potential depression or the user including such a transition in the atomic data input.

ci: picongpu