fix folder creation
fix #5401

The method `createDirectory()` and `createDirectoryWithPermissions()` was not clearly mention that only one MPI rank is allowed to call this method.
The original implementation guarded the permission setting method by using the global rank zero only but in cases where the IO rank was not rank zero the permission was not set and since the refactoring #5094 also the output folder was not created.