add MallocMC to PMacc using fetchContent
The current pattern fails if there are multiple passes of configure and
`mallocMC` was not initially installed.
This happens because in the first pass `add_subdirectory` creates the
package config and in later passes `findPackage` finds this and fails as
`mallocMC` wasnt actually installed.
Switched to simply using `FetchContent` with `FIND_PACKAGE_ARGS` which
first checks with `findPackage` and then falls back to the source dir
which is exactly what we want.
https://cmake.org/cmake/help/latest/guide/using-dependencies/index.html#fetchcontent-and-find-package-integration